double instead of float to reduce rounding errors

+++ lib/printer-emulator/atari_1029.h
@@ -32,8 +32,8 @@ protected:
     void clear_mode(uint16_t m);
     void reset_cmd();
     uint8_t epson_font_lookup(uint16_t code);
-    float epson_font_width(uint16_t code);
-    void epson_set_font(uint8_t F, float w);
+    double epson_font_width(uint16_t code);
+    void epson_set_font(uint8_t F, double w);
 
     virtual void pdf_clear_modes() override;
     void pdf_handle_char(uint8_t c, uint8_t aux1, uint8_t aux2) override;
